Rename hlb package
