+++ cmd/browse_issues.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/mpppk/hlb/hlb"
+	"github.com/mpppk/hlb/hlblib"
 	"github.com/mpppk/hlb/etc"
 	"github.com/skratchdot/open-golang/open"
 	"fmt"
@@ -19,9 +19,9 @@ var browseissuesCmd = &cobra.Command{
 			fmt.Println("Too many issue IDs")
 		}
 
-		base, err := hlb.NewCmdBase()
+		base, err := hlblib.NewCmdBase()
 		etc.PanicIfErrorExist(err)
-		sw := hlb.ClientWrapper{Base: base}
+		sw := hlblib.ClientWrapper{Base: base}
 
 		if len(args) == 0 {
 			url, err := sw.GetIssuesURL()
